+++ include/solarus/entities/Hero.h
@@ -399,7 +399,7 @@ class Hero: public Entity {
     int ice_movement_direction8;           /**< wanted movement direction a while ago */
     Point ground_dxy;                      /**< additional movement with special ground (hole or ice) */
 
-    //behavior
+    // behavior
     int push_delay;                        /**< delay before going into pushing state */
 
 };